+++ shared/src/commonMain/kotlin/ui/navigation/WalletNavigation.kt
@@ -262,6 +262,10 @@ private fun WalletNavHost(
                 },
                 onDismiss = {
                     walletMain.credentialValidityService.removeRefreshRequest(item)
+                },
+                onNeverShowAgain = {
+                    processedItemIds = processedItemIds + item.storeEntryId
+                    walletMain.credentialValidityService.suppressRefreshRequest(item)
                 }
             )
         }